requirements.txt
pystray
pillow
pyvda